requirements.txt
apache-airflow==2.8.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
torch==2.1.2
jax[cpu]==0.4.23
flax==0.7.5
fastapi==0.109.0
uvicorn==0.27.0
mlflow==2.10.0
evidently==0.4.13
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyarrow==14.0.2
pydantic==2.5.3
